requirements.txt
streamlit
langchain
python-dotenv
langchain-experimental
openai
pandas
tabulate